==> 10_nn_data_impute.ipynb <==
# Modify the file 00_setup.py to define input/output file paths on your system
# The information in 00_setup.py will be used across notebooks
from importlib.machinery import SourceFileLoader
setup = SourceFileLoader("setup", "./00_setup.py").load_module()

# Data Imputation and Scaling for Neural Networks
Convert features so they can be used in a neural network (or GNN).  
* Impute missing features
* Scale values
* Set an "alternative" NAICS for encoding
  * For NAICS not in the training data, use the highest-volume NAICS within the sector

For continuous features, I do quantile encoding then min/max scaling to -1, 1.  For binary fields (and fields with a small number of levels) I just min/max rescale

*This script takes about 5 minutes on my MacBook Air*

import pandas as pd
import numpy as np

from pathlib import Path
import importlib, pickle
from sklearn import model_selection
from sklearn.preprocessing import LabelEncoder

# Imputer object for easy dataset conversion to GNN friendly format
from sba_gnn.sba_gnn import sg_imputer 
from sba_gnn.sba_gnn.sg_imputer import GNNImputer

## Input training data, set alternate NAICS

Import the data, and set an alternate NAICS for test and validation, such that NAICS that are not in the training data are set to the most common NAICS in the same sector in the training data

##### Import train data

sba_loans = pd.read_parquet(Path(setup.temp_path).joinpath('01_DATA_transformed.parquet'))

## Get alternative NAICS (test and validation)
Map unseen NAICS using hierarchical information.  Use the highest-volume code in a grouping

def get_common_naics(data, group_var= 'NAICS_sector', alt_var = 'NAICS_alt'):
    """Get an alternative NAICS as the most common value in a larger group"""
    grouped = data.groupby(group_var)
    alt_naics = pd.concat([g[['NAICS']].value_counts().head(1).reset_index() \
                              for n, g in grouped], keys=[n for n, g in grouped]) \
        .reset_index(level=0) \
        .reset_index(drop = True) \
        .rename(columns={'level_0':group_var, 'NAICS':alt_var}) \
        [[group_var, alt_var]] 
    return alt_naics

def remap_naics(data, group_var= 'NAICS_sector', alt_var = 'NAICS_alt'):
    """Map unseen test/validation codes to same-group codes"""
    train_data = data[data['dset'] == 'train']
    oth_data = data[data['dset'] != 'train']
    
    # Do not modify the train data 
    naics_alt_train = train_data[['NAICS', 'LoanNr_ChkDgt']].rename(columns={'NAICS':alt_var})
    
    # Get alternative NAICS 
    grp_alt = get_common_naics(train_data, group_var, alt_var)
    
    # Identify seen/unseen codes
    train_naics = train_data[['NAICS']].drop_duplicates()
    naics_alt_merge = oth_data[['LoanNr_ChkDgt', 'NAICS', group_var]] \
        .merge(train_naics, how='left', on='NAICS', indicator = True)
    
    # Don't map seen codes
    naics_alt_2 = naics_alt_merge[naics_alt_merge['_merge'] == 'both'] \
        .drop(columns=['_merge', group_var], errors='ignore') \
        .rename(columns={'NAICS':alt_var})
    
    # Map unseen codes
    naics_alt_3 = naics_alt_merge[naics_alt_merge['_merge'] == 'left_only'] \
        .drop(columns=['_merge']) \
        .merge(grp_alt, how='left', on=group_var) \
        .drop(columns=[group_var])
    
    # If no match, use same unseen code
    naics_alt_3[alt_var] = naics_alt_3[alt_var].fillna(naics_alt_3['NAICS'])
    naics_alt_3.drop(columns='NAICS', inplace=True)
    
    naics_alt = pd.concat([naics_alt_train, naics_alt_2, naics_alt_3])[['LoanNr_ChkDgt', alt_var]]
    
    return data.drop(columns=[alt_var], errors='ignore') \
        .merge(naics_alt, on='LoanNr_ChkDgt')        

##### Sector map

sba_loans = remap_naics(sba_loans)

sba_loans.shape

(sba_loans['NAICS_alt'] == sba_loans['NAICS']).value_counts()

sba_loans['NAICS_alt'].isna().value_counts()

##### Subsector map

sba_loans['naics3'] = sba_loans['NAICS'].str[0:3]

sba_loans = remap_naics(sba_loans, group_var='naics3', alt_var = 'NAICS_alt3')

(sba_loans['NAICS_alt3'] == sba_loans['NAICS']).value_counts()

sba_loans['NAICS_alt3'].isna().value_counts()

##### Industry group map

sba_loans['naics4'] = sba_loans['NAICS'].str[0:4]

sba_loans = remap_naics(sba_loans, group_var='naics4', alt_var = 'NAICS_alt4')

(sba_loans['NAICS_alt4'] == sba_loans['NAICS']).value_counts()

sba_loans['NAICS_alt4'].isna().value_counts()

## Imputer

train_df = sba_loans[sba_loans['dset'] == 'train']
test_val_df = sba_loans[sba_loans['dset'] != 'train']

For a neural network, I need to handle missings and also scale features.  For missing values, I will do a simple median fill for all, but add missing indicators.  

To scale the data, I will do a quantile transform for features with > 5 levels.  This is to avoid scaling binary features.  Then I do a Min/Max scaling on all features, so they are in the [0,1] range

I define a class to do this in sg_imputer.  This class wraps several scikit-learn imputers/scalers so that I can easily fit the objects and then transform the data.  In this section, I show some descriptives to justify my imputation/scaling choices, and also fit the scaler

train_df = sba_loans[sba_loans['dset'] == 'train']

##### Descriptives - explore missingness

# Same information, more compact
pd.concat([pd.DataFrame([train_df[c].isna().sum()]) for c in setup.predictor_features],
         keys=setup.predictor_features).set_axis(['count_missing'], axis=1)

Most fields aren't missing too much.  However I must handle those that are.  I will use a median fill, especially since many of the fields with missingness seem to be binary

##### Explore scaling

# Level counts
feature_level_df = pd.concat([pd.DataFrame([train_df[c].value_counts().count()]) for c in setup.predictor_features],
         keys=setup.predictor_features)
feature_level_df

A number of features are binary.  I don't need to scale these.  The others must be scaled.  

What I will do is scale features with more than a threshold number of levels using a quantile scaler.  Then I'll use a MinMax scaler on everything to be safe (as a 2 level feature could have values other than 0,1).

##### Create imputer to do the median fill and scaling
This imputer fills missing values, adding missingness indicator features.  Then it quantile scales features with over 5 levels, and MinMax scales the rest.

importlib.reload(sg_imputer)
from sba_gnn.sba_gnn.sg_imputer import GNNImputer

imputer = GNNImputer(features = setup.predictor_features,
                    naics_features = ['NAICS', 'NAICS_alt', 'NAICS_alt3', 'NAICS_alt4'])                               

train_out = imputer.fit_transform(train_df)

train_out.describe()

with open(Path(setup.temp_path).joinpath('10_DATA_imputer.pkl'), 'wb') as fout:
    pickle.dump(imputer, fout)

with open(Path(setup.temp_path).joinpath('10_DATA_features.pkl'), 'wb') as fout:
    pickle.dump(imputer.features_out, fout)

naics_max_levels = imputer.get_naics_encoder_levels()
print(f'NAICS encoder max num: {naics_max_levels}')
with open(Path(setup.temp_path).joinpath('10_DATA_naics_max_levels.pkl'), 'wb') as fout:
    pickle.dump(naics_max_levels, fout)

## Transform Test, Validation Data

test_df = sba_loans[sba_loans['dset'] == 'test']
val_df = sba_loans[sba_loans['dset'] == 'val']

test_out = imputer.transform(test_df)

val_out = imputer.transform(val_df)

pd.concat([test_out[['DisbursementGross', 'NoEmp']], test_df[['DisbursementGross', 'NoEmp']]],
          axis=1).corr(method='spearman')

val_out.describe()

test_out.describe()

test_out[['NAICS', 'NAICS_alt']].corr()

val_out[['NAICS', 'NAICS_alt']].corr()

train_out[['NAICS', 'NAICS_alt']].corr()

## Combine data
Combine transformed datasets, with key information.  Include the business ID, target status, NAICS features, as well as transformed features from above. Save this data for later use (especially the indices which indicate train/test statustrain_df

naics_features = ['LoanNr_ChkDgt', 'target', 'dset_naics_holdout', 'NAICS', 'NAICS_sector', 
                  'menc_NAICS', 'cenc_NAICS','NAICS_alt', 'NAICS_alt3', 'NAICS_alt4'] +  \
    [c for c in train_df.columns if c.startswith('NS__')]
print(naics_features)

comb_naics = pd.concat([train_df[naics_features], 
                        test_df[naics_features], 
                        val_df[naics_features]], axis=0, keys=['train', 'test', 'val']) \
    .rename(columns={'NAICS':'NAICS_orig', 'NAICS_alt':'NAICS_alt_orig', 
                     'NAICS_alt3':'NAICS_alt3_orig', 'NAICS_alt4':'NAICS_alt4_orig'}) \
    .sort_index()

comb_df = pd.concat([comb_naics,
                     pd.concat([train_out, test_out, val_out], 
                               axis=0, keys=['train', 'test', 'val']).sort_index()],
                    axis=1) \
    .reset_index(level=0) \
    .rename(columns={'level_0':'dset'}, errors='ignore')
print(comb_df.shape)

print(comb_df.columns)

comb_df.to_parquet(Path(setup.temp_path).joinpath('10_DATA_combined_scaled_all.parquet'))
