==> 32_dgi_label_data.ipynb <==
# Modify the file 00_setup.py to define input/output file paths on your system
# The information in 00_setup.py will be used across notebooks
from importlib.machinery import SourceFileLoader
setup = SourceFileLoader("setup", "./00_setup.py").load_module()

# GNN - Unsup Model Data, With Label

Create Stellargraph objects for unsupervised GNN runs, including original features, plus the target as a feature.

For the unseen / test data, randomly set the target value in the feature

*This script takes about 15 minutes on my MacBook Air*

import pandas as pd
import numpy as np

from pathlib import Path
import importlib
import pickle, re

from sba_gnn.sba_gnn import sg_gnn

import stellargraph as sg
from stellargraph import StellarGraph

## Import Data

## NN scaled input data
comb_df = pd.read_parquet(Path(setup.temp_path).joinpath('20_DATA_combined_scaled_all.parquet'))

with open(Path(setup.temp_path).joinpath('20_DATA_features.pkl'), 'rb') as fin:
    imputed_features = pickle.load(fin)

num_feat =  [c for c in imputed_features if 'NAICS' not in c]
print(num_feat)

###### Data used to create previous graphs

edges_all = pd.read_parquet(Path(setup.temp_path).joinpath('30_DATA_graph_edges.parquet'))

node_features_business = pd.read_parquet(Path(setup.temp_path).joinpath('30_DATA_graph_node_features_loans.parquet'))

node_features_naics = pd.read_parquet(Path(setup.temp_path).joinpath('30_DATA_graph_node_features_naics.parquet'))

## Add target to node features
Add target data to the features for the train and validation data.  For test, randomly set

targ_feat_train_val = comb_df[comb_df['dset'].isin(['train', 'val'])] \
    [['LoanNr_ChkDgt', 'target', 'dset']] \
    .set_index('LoanNr_ChkDgt') \
    .rename(columns={'target':'feature_target'})

base_target_rate = targ_feat_train_val[targ_feat_train_val['dset']=='train']['feature_target'].mean()
base_target_rate

targ_feat_test = comb_df[comb_df['dset'] == 'test'] \
    [['LoanNr_ChkDgt', 'dset']] \
    .set_index('LoanNr_ChkDgt') 

np.random.seed(1234)
targ_feat_test['feature_target'] = np.random.choice([1, 0], 
                                                    p = [base_target_rate, 1-base_target_rate],
                                                    size = len(targ_feat_test))

targ_feat = pd.concat([targ_feat_train_val, targ_feat_test]).sort_index()

node_features_business = pd.concat([node_features_business.sort_index(),
                                    targ_feat[['feature_target']]], axis=1)
node_features_business.to_parquet(Path(setup.temp_path).joinpath('32_DATA_graph_node_features_loans.parquet'))

node_features_business.head(3)

## Stellargraph - train and validation
Graph without test nodes (make sure set-aside NAICS not included)

# Get only relevant edges and node data
edges_train_val, features_business_train_val, features_naics_train_val = \
    sg_gnn.limit_data(edges_all, node_features_business, node_features_naics.drop(columns='NAICS_orig'),
                     ['train', 'val'])

print(f'Train+val graph data')
print(f'business features, start rows {node_features_business.shape[0]}, end {features_business_train_val.shape[0]}')
print(f'naics features, start rows {node_features_naics.shape[0]}, end {features_naics_train_val.shape[0]}')
print(f'edges, start rows {edges_all.shape[0]}, end {edges_train_val.shape[0]}')

sba_graph_train_val = StellarGraph({'LoanNr_ChkDgt':features_business_train_val.drop(columns='NAICS_orig'),
                                    'NAICS': features_naics_train_val},
                                   edges_train_val, 
                                   source_column="source", target_column="target",
                                   edge_type_column="type")

print(sba_graph_train_val.info())

# Save stellargraph object
with open(Path(setup.temp_path).joinpath('32_DATA_stellargraph_train_val.pkl'), 'wb') as fout:
      pickle.dump(sba_graph_train_val, fout)

## Stellargraph - all nodes

sba_graph_all = StellarGraph({'LoanNr_ChkDgt':node_features_business.drop(columns=['dset', 'NAICS_orig']) ,
                              'NAICS': node_features_naics.drop(columns='NAICS_orig')},
                             edges_all.drop(columns=['dset']).drop_duplicates(),
                             source_column="source", target_column="target",
                             edge_type_column="type")

print(sba_graph_all.info())

# Save stellargraph object
with open(Path(setup.temp_path).joinpath('32_DATA_stellargraph_all.pkl'), 'wb') as fout:
      pickle.dump(sba_graph_all, fout)
