==> 80_perf_summary.ipynb <==
# Modify the file 00_setup.py to define input/output file paths on your system
# The information in 00_setup.py will be used across notebooks
from importlib.machinery import SourceFileLoader
setup = SourceFileLoader("setup", "./00_setup.py").load_module()

# Performance Summary
Combine performance of models tested, and summarize

Also combine all metrics for easy comparisons

*This script takes about 1 minutes on my MacBook Air*

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from pathlib import Path

from sba_gnn.sba_gnn import sg_plot 

sg_plot.plot_defaults()

## File Dictionaries

##### XGB basic tests

dict_pred_1_xgb = {
    'xgb base': Path(setup.temp_path).joinpath('03_DATA_combined_predictions.parquet'),
    'xgb mean':  Path(setup.temp_path).joinpath('04_DATA_combined_predictions.parquet'),
    'xgb one hot':  Path(setup.temp_path).joinpath('05_DATA_combined_predictions.parquet'),
    'xgb numeric naics':  Path(setup.temp_path).joinpath('06_DATA_combined_predictions.parquet'),
    'xgb mean':  Path(setup.temp_path).joinpath('07_DATA_combined_predictions.parquet'),
    'xgb hier': Path(setup.temp_path).joinpath('10_DATA_combined_predictions.parquet')
}

##### Neural network basic

dict_pred_2_nn_basic = {
    'nn base': Path(setup.temp_path).joinpath('21_DATA_predictions.parquet'),
    'nn emb': Path(setup.temp_path).joinpath('22_DATA_predictions.parquet'),
    'nn menc': Path(setup.temp_path).joinpath('23_DATA_predictions.parquet'),
    'nn hier': Path(setup.temp_path).joinpath('24_DATA_predictions.parquet'),
    'nn emb remap': Path(setup.temp_path).joinpath('25_DATA_predictions.parquet'),
}

##### XGB + embeddings

dict_pred_3_emb = {
    'xgb emb nn':  Path(setup.temp_path).joinpath('40_DATA_combined_predictions.parquet'),
    'xgb dgi base':  Path(setup.temp_path).joinpath('50_DATA_combined_predictions.parquet'),
    'xgb dgi base+label':  Path(setup.temp_path).joinpath('52_DATA_combined_predictions.parquet')
}

##### Clustering - Mean Encoding

dict_pred_4_clus_menc = {
    'xgb clus menc nn':  Path(setup.temp_path).joinpath('62_DATA_combined_predictions.parquet'),
    'xgb clus dgi hier':  Path(setup.temp_path).joinpath('65_DATA_combined_predictions.parquet')
}

##### Clustering - One Hot Encoding

dict_pred_5_oh = {
    'xgb clus oh nn':  Path(setup.temp_path).joinpath('61_DATA_combined_predictions.parquet'),
    'xgb clus oh dgi base':  Path(setup.temp_path).joinpath('64_DATA_combined_predictions.parquet'),
    'xgb clus oh dgi base+label':  Path(setup.temp_path).joinpath('71_DATA_combined_predictions.parquet')
}

## Loop Dictionaries
Some scripts have multiple models - pull these in a little differently

dict_loop = {
        'xgb hier vary start':  Path(setup.temp_path).joinpath('11_DATA_combined_predictions.parquet'),
        'xgb hier pair':  Path(setup.temp_path).joinpath('12_DATA_combined_predictions.parquet'),
        'xgb dgi hier v start':  Path(setup.temp_path).joinpath('66_DATA_combined_predictions.parquet'),
        'xgb dgi hier v start (alt)':  Path(setup.temp_path).joinpath('67_DATA_combined_predictions.parquet'),
        'xgb dgi hier pair':  Path(setup.temp_path).joinpath('68_DATA_combined_predictions.parquet')
}

## Combine predictions

##### Get dictionary of all files

all_dict_list = [dict_pred_1_xgb, dict_pred_2_nn_basic, 
                 dict_pred_3_emb, dict_pred_4_clus_menc, dict_pred_5_oh, dict_loop]

dict_pred= {k: v for d in all_dict_list for k, v in d.items()}

# Verify these all exist
all([Path(v).exists() for k, v in dict_pred.items()])

##### Standard metrics
Get filenames for standard metrics calculated in all scripts. These all have similar file naming

dict_metrics = {k:Path(setup.temp_path).joinpath(v.name[0:2] + '_REPORT_metrics.csv') \
                for k, v in dict_pred.items()}

# Verify these all exist
all([Path(v).exists() for k, v in dict_metrics.items()])

## Combine standard metrics

all_metrics = pd.concat([pd.read_csv(v) for k, v in dict_metrics.items()],
                        keys = dict_metrics.keys()) \
    .reset_index(level=0) \
    .drop(columns=['Unnamed: 0'], errors='ignore') \
    .rename(columns={'level_0':'model'})

all_metrics['script_num'] = all_metrics['model'].apply(lambda x: dict_metrics[x].name[0:2])

##### Modify loop field names

all_metrics[all_metrics['model'].isin(dict_loop.keys())]

all_metrics['model'] = np.where(all_metrics['model'].isin(dict_loop.keys()),
                                all_metrics['model'] + ':' + all_metrics['start_naics'],
                                all_metrics['model'])

all_metrics.tail(3)

all_metrics.to_csv(Path(setup.temp_path).joinpath('80_REPORT_metrics_combined.csv'), index=False)

## Summary Data
For test dataset only, show AUC and f1 scores for test dataset, along with f1 scores for holdout and low-volume NAICS

overall_stats =  all_metrics[(all_metrics['dset'] == 'test') & \
                             (all_metrics['dset_naics_holdout'] == 0)] \
    [['script_num', 'model', 'roc_auc_score', 'average_precision_score', 'f1_score', ]]

holdout_stats = all_metrics[(all_metrics['dset'] == 'test') & \
                            (all_metrics['dset_naics_holdout'] == 1)] \
    [['model', 'roc_auc_score', 'average_precision_score', 'f1_score']] \
    .rename(columns={'f1_score':'f1_score_ho', 'average_precision_score':'ap_ho',
                    'roc_auc_score':'roc_ho'})

overall_stats = overall_stats.merge(holdout_stats, on='model')

##### Selected stats

overall_stats.sort_values('average_precision_score', ascending=False).head(12)

overall_stats.sort_values('ap_ho', ascending=False).head(12)

overall_stats[overall_stats['model'].isin(list(dict_pred_2_nn_basic.keys()))] \
    .sort_values('average_precision_score', ascending=False)

overall_stats[overall_stats['model'].isin(list(dict_pred_3_emb.keys()))] \
    .sort_values('average_precision_score', ascending=False)

overall_stats[overall_stats['model'].isin(dict_pred_4_clus_menc.keys())] \
    .sort_values('average_precision_score', ascending=False)

overall_stats[overall_stats['model'].isin(list(dict_pred_5_oh.keys()) + \
                                          ['xgb one hot'])] \
    .sort_values('average_precision_score', ascending=False)

overall_stats.to_csv(Path(setup.temp_path).joinpath('80_REPORT_summary_stats.csv'),
                    index=False)
