==> 81_anova.ipynb <==
# Modify the file 00_setup.py to define input/output file paths on your system
# The information in 00_setup.py will be used across notebooks
from importlib.machinery import SourceFileLoader
setup = SourceFileLoader("setup", "./00_setup.py").load_module()

# ANOVA, Regression, etc.
Various ANOVA and related analyses

import numpy as np
import pandas as pd
from pathlib import Path
import importlib, os

import statsmodels.api as sm
from statsmodels.formula.api import ols
from scipy import stats

import sklearn as sk
sk.__version__

## Import Data

sba_loans = pd.read_parquet(Path(setup.temp_path).joinpath('01_DATA_transformed.parquet'))

train_df = sba_loans[sba_loans['dset'] == 'train']

##### NAICS sector (one hot encodings) info

one_hot_feat= [c for c in sba_loans.columns if c.startswith('NS__')]
one_hot_feat_len = len(one_hot_feat)
one_hot_feat_len

one_hot_feat

##### NAICS info

naics_info = pd.read_parquet(Path(setup.temp_path).joinpath('60_DATA_naics_summary_stats.parquet')) 

naics_map = pd.read_parquet(Path(setup.temp_path).joinpath('60_DATA_naics_map.parquet'))

naics_info.sample(2)

##### Neural Network Embeddings
With cluster and NAICS info appended

emb_nn_clus = pd.read_parquet(Path(setup.temp_path).joinpath('60_DATA_embeddings_tsne_naics.parquet'))
emb_nn_feat = [c for c in emb_nn_clus.columns if c.startswith('emb_')]

# Other size clusters, to get correct size
nn_cluster_labels_all = pd.read_parquet(Path(setup.temp_path).joinpath('60_DATA_loop_cluster_labels.parquet'))

##### DGI embeddings

emb_dgi_clus = pd.read_parquet(Path(setup.temp_path).joinpath('63_DATA_embeddings_tsne_naics.parquet'))
emb_dgi_feat = [c for c in emb_dgi_clus.columns if c.startswith('emb_')]

## Functions

##### ANOVA with variation explained

def anova_var(data, eqn):
    """Perform one way ANOVA, return variance explained"""
    lm = ols(eqn, data=data).fit()
    res = sm.stats.anova_lm(lm, typ=1)
    res = pd.concat([res, 
                     (res['sum_sq'].transform(lambda x: x/x.sum())).rename('var_f')],
                    axis=1)
    return res

## NAICS sector - Baseline
ANOVA results in the training data, looking at variation explained by NAICS groupings

##### Stats oneway summary

sector_groups = train_df.groupby('NAICS_sector')
sectors_list = [g['target'].to_numpy() for n, g in sector_groups]

f_sector, p_sector = stats.f_oneway(*sectors_list)

print(f'sector oneway f: {f_sector}, p: {p_sector}')

naics_groups = train_df.groupby('NAICS')
naics_list = [g['target'].to_numpy() for n, g in naics_groups]

f_naics, p_naics = stats.f_oneway(*naics_list)

print(f'naics oneway f: {f_naics}, p: {p_naics}')

##### Residuals analysis - sector

res = anova_var(train_df, 'target ~ C(NAICS_sector)')

res

##### NAICS level 
How much of the mean target by NAICS is accounted for by sector?

naics_info_train = naics_info[naics_info['train_count'] > 0]

anova_var(naics_info_train, 'train_target_mean ~ C(NAICS_sector)')

## NN Embeddings

##### NAICS sector ANOVA

# OLS strings for each
emb_str = [ ef+ ' ~ C(NAICS_sector)' for ef in emb_nn_feat]

# All ANOVAs
anova_nn_sector = pd.concat([anova_var(emb_nn_clus, s) for s in emb_str],
                     keys=emb_nn_feat) \
    .reset_index()

anova_nn_sector[anova_nn_sector['level_1'] == 'C(NAICS_sector)']

##### NAICS level ANOVA (mean target)

naics_info_train_nn_cluster = naics_info[naics_info['train_count'] > 0] \
    .merge(emb_nn_clus.drop(columns='NAICS').rename(columns={'NAICS_orig':'NAICS'})[['NAICS', 'cluster']],
           how='left', on='NAICS')

anova_var(naics_info_train_nn_cluster, 'train_target_mean ~ C(cluster)')

##### Loan level ANOVA (raw target)

train_nn_cluster = train_df \
    .merge(emb_nn_clus.drop(columns='NAICS').rename(columns={'NAICS_orig':'NAICS'})[['NAICS', 'cluster']],
           how='left', on='NAICS')

anova_var(train_nn_cluster, 'target ~ C(cluster)')

## DGI Embeddings

##### NAICS sector ANOVA

# OLS strings for each
emb_dgi_str = [ ef+ ' ~ C(NAICS_sector)' for ef in emb_dgi_feat]

# All ANOVAs
anova_dgi_sector = pd.concat([anova_var(emb_dgi_clus, s) for s in emb_dgi_str],
                     keys=emb_nn_feat) \
    .reset_index()

anova_dgi_sector[anova_nn_sector['level_1'] == 'C(NAICS_sector)']

##### NAICS level ANOVA (mean target)

naics_info_train_dgi_cluster = naics_info[naics_info['train_count'] > 0] \
    .merge(emb_dgi_clus.rename(columns={'NAICS_orig':'NAICS'})[['NAICS', 'cluster']],
           how='left', on='NAICS')

anova_var(naics_info_train_dgi_cluster, 'train_target_mean ~ C(cluster)')

##### Loan level ANOVA (raw target)

train_dgi_cluster = train_df \
    .merge(emb_dgi_clus.rename(columns={'NAICS_orig':'NAICS'})[['NAICS', 'cluster']],
           how='left', on='NAICS')

anova_var(train_dgi_cluster, 'target ~ C(cluster)')

## Logistic Regression
Very simple models to analyze information in high-level groupings (Sector and from embeddings)

Use same-size one hot features for each group, train data only.

##### Get fixed-level NN clusters 
The NN clusters are only for training; get one_hot_feat_len of these (unlike earlier scripts where I used 

sel_labels_nn = emb_nn_clus[['NAICS_orig']] \
    .merge(nn_cluster_labels_all[nn_cluster_labels_all['n_clusters'] == one_hot_feat_len] \
               .drop(columns='n_clusters'),
           left_index=True, right_index = True)
sel_labels_nn.shape

# DGI clusters - limit to training match
emb_dgi_train = emb_dgi_clus \
    .rename(columns={'cluster':'cluster_dgi','NAICS_orig':'NAICS'}) \
    .merge(train_df[['NAICS']].drop_duplicates(), on='NAICS') 

all_naics_clus = emb_dgi_train \
    .merge(sel_labels_nn.rename(columns={'NAICS_orig':'NAICS', 'label':'cluster_nn'}), on='NAICS')
all_naics_clus[['cluster_dgi', 'cluster_nn']] = all_naics_clus[['cluster_dgi', 'cluster_nn']] \
    .apply(lambda x: x.astype('category'))

lm_df = train_df[['LoanNr_ChkDgt', 'NAICS','dset', 'NAICS_sector', 'target']] \
     .merge(all_naics_clus[['NAICS', 'cluster_dgi', 'cluster_nn']],
                          how='left', on='NAICS') \
    .merge(naics_info[['NAICS', 'NAICS_sector_sel']], how='left', on='NAICS')

lm_df['NAICS_sector_sel'] = np.where(lm_df['NAICS_sector_sel'] == 1,lm_df['NAICS_sector'],
                                     'other')
#lm_df['NAICS_sector_sel'] = lm_df['NAICS_sector_sel'].astype('category')

##### Function to do the logistic regression

def log_reg(data, feature = ['NAICS_sector_sel']):
    X = pd.get_dummies(data[feature], dtype='float', drop_first=True)
    Y = data['target']
    Xi = sm.add_constant(X)
    sm_logreg = sm.Logit(Y, Xi)

    res = sm_logreg.fit_regularized(method='l1')

    return res
                            

##### Single group models

res_naics_sector = log_reg(lm_df)
print(res_naics_sector.summary())

res_cluster_dgi = log_reg(lm_df, ['cluster_dgi'])
print(res_cluster_dgi.summary())

res_cluster_nn = log_reg(lm_df, ['cluster_nn'])
print(res_cluster_nn.summary())

##### Two-way models

res_2_sector_nn = log_reg(lm_df, ['NAICS_sector_sel', 'cluster_nn'])
print(res_2_sector_nn.summary())

res_2_sector_dgi = log_reg(lm_df, ['NAICS_sector_sel', 'cluster_dgi'])
print(res_2_sector_dgi.summary())

res_2_nn_dgi = log_reg(lm_df, ['cluster_nn', 'cluster_dgi'])
print(res_2_nn_dgi.summary())

##### 3 Way

res_3 = log_reg(lm_df, ['cluster_nn', 'cluster_dgi', 'NAICS_sector_sel'])
print(res_3.summary())

##### Gather Metrics

def get_metric_ser(res):
    return pd.Series([res.aic, res.bic, res.prsquared],
                     index=['aic', 'bic', 'prsquared'])

res_list = [res_naics_sector, res_cluster_dgi, res_cluster_nn, 
            res_2_sector_dgi, res_2_sector_nn, res_2_nn_dgi, res_3]
res_name_list = ['sector', 'dgi', 'nn',  'sector + dgi', 'sector + nn',
                 'dgi + nn', 'sector + dgi + nn']
logit_metrics = pd.concat([get_metric_ser(x).to_frame().transpose() for x in res_list],
                          keys=res_name_list, axis=0) \
    .reset_index(level=1, drop=True) \
    .reset_index() \
    .rename(columns={'index':'model'})
logit_metrics

logit_metrics.to_csv(Path(setup.temp_path).joinpath('81_REPORT_logit_metrics.csv'), index=False)
